requirements.txt
waitress
Flask
Python
mysql-connector-python
ParseHub
Task Scheduler
